Move holidayStats into own service file
