+++ src/pages/Profile/container.js
@@ -3,7 +3,8 @@ import { PropTypes as PT } from 'prop-types';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { getUser } from '../../reducers';
-import { getEvents, getHolidayStats } from '../../services/eventService';
+import { getEvents } from '../../services/eventService';
+import { getHolidayStats } from '../../services/holidayService';
 import { isEmpty } from 'lodash';
 import { getContractsByEmployeeId } from '../../services/contractService';
 import eventTypes from '../../utilities/eventTypes';
