+++ src/services/holidayService.js
@@ -1,0 +1,5 @@
+import axios from '../utilities/AxiosInstance';
+
+export const getHolidayStats = employeeId => {
+  return axios.get(`/Event/findEmployeeHolidayStats/${employeeId}`);
+};
